refactor: 객체명 변경 및 패키지 이동 (#63)
