+++ src/main/java/com/beside/archivist/exception/users/UserNotFoundException.java
@@ -1,6 +1,6 @@
 package com.beside.archivist.exception.users;
 
-import com.beside.archivist.exception.ExceptionCode;
+import com.beside.archivist.exception.common.ExceptionCode;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
